minor fixwa

+++ main/views.py
@@ -8,46 +8,52 @@ from datetime import date
 def home(response):
    return render(response, "home.html")
 
-@login_required
+#@login_required
 def plan(response):
-   d = Day.objects.filter(user=response.user)
-   #print(d)
-   do = [choice[0] for choice in Day.daysofweek.choices]
-   d = sorted(d, key=lambda d: do.index(d.day))
-   #print( d)
-   #print(d[0].day)
-   #print(d[0].todo)
-   return render(response, "plan.html",{"days":d})
+    if response.user.is_authenticated:
+        d = Day.objects.filter(user=response.user)
+        #print(d)
+        do = [choice[0] for choice in Day.daysofweek.choices]
+        d = sorted(d, key=lambda d: do.index(d.day))
+        #print( d)
+        #print(d[0].day)
+        #print(d[0].todo)
+        return render(response, "plan.html",{"days":d})
+    else:
+        return render(response,"plan.html")
 
-@login_required
+#@login_required
 def edit(response):
-   if response.method == "POST":
-        if "delete_workout" in response.POST:
-            workout_id = response.POST.get("workout_id")
-            workout = get_object_or_404(Workout, id=workout_id, tday__user=response.user)
-            workout.delete()
-            return redirect("edit")
+    if response.user.is_authenticated:
+        if response.method == "POST":
+                if "delete_workout" in response.POST:
+                    workout_id = response.POST.get("workout_id")
+                    workout = get_object_or_404(Workout, id=workout_id, tday__user=response.user)
+                    workout.delete()
+                    return redirect("edit")
 
-        day_id = response.POST.get("day_id")
-        d = get_object_or_404(Day, id=day_id, user=response.user)
+                day_id = response.POST.get("day_id")
+                d = get_object_or_404(Day, id=day_id, user=response.user)
 
-        if "save_todo" in response.POST:
-            d.todo = response.POST.get("todo", d.todo)
-            d.save()
+                if "save_todo" in response.POST:
+                    d.todo = response.POST.get("todo", d.todo)
+                    d.save()
 
-        elif "save_workout" in response.POST:
-            form = addW(response.POST)
-            if form.is_valid():
-                workout = form.save(commit=False)
-                workout.tday = d
-                workout.save()
+                elif "save_workout" in response.POST:
+                    form = addW(response.POST)
+                    if form.is_valid():
+                        workout = form.save(commit=False)
+                        workout.tday = d
+                        workout.save()
 
-        return redirect("edit")   
-   form = addW()
-   d = Day.objects.filter(user=response.user)
-   do = [choice[0] for choice in Day.daysofweek.choices]
-   d = sorted(d, key=lambda d: do.index(d.day))
-   return render(response, "edit.html",{"days":d,"form":form} )
+                return redirect("edit")   
+        form = addW()
+        d = Day.objects.filter(user=response.user)
+        do = [choice[0] for choice in Day.daysofweek.choices]
+        d = sorted(d, key=lambda d: do.index(d.day))
+        return render(response, "edit.html",{"days":d,"form":form} )
+    else:
+       return render(response, "edit.html")
 
 weekdaynum = {
         0: "MON",
@@ -59,8 +65,11 @@ weekdaynum = {
         6: "SUN",
     }
 
-@login_required
+#@login_required
 def today(response):
-    todayc = weekdaynum[date.today().weekday()]
-    d = get_object_or_404(Day, user=response.user, day=todayc)
-    return render(response,"today.html",{"d":d, "todayd":date.today()})
+    if response.user.is_authenticated:
+        todayc = weekdaynum[date.today().weekday()]
+        d = get_object_or_404(Day, user=response.user, day=todayc)
+        return render(response,"today.html",{"d":d, "todayd":date.today()})
+    else:
+        return render(response,"today.html")
